requirements.txt
groq==0.9.0
chromadb==0.5.0
sentence-transformers==3.0.1
pdfplumber==0.11.0
python-docx==1.1.2
streamlit==1.36.0
python-dotenv==1.0.1
pandas==2.2.2